requirements.txt
flask==2.3.3
torch==2.0.1
transformers==4.30.2
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
scikit-image==0.21.0
opencv-python==4.8.0.74
requests==2.31.0
googletrans==4.0.0-rc1
